backend/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
beautifulsoup4==4.13.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django
django-cors-headers==4.6.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
idna==3.10
numpy
packaging==24.2
pandas==2.2.3
pillow==11.1.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
redis==5.2.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
whitenoise==6.9.0
XlsxWriter==3.2.3
